requirements.txt
django==3.2.7
djangorestframework==3.12.4
djoser==2.1.0
djangorestframework_simplejwt==4.8.0
djangochannelsrestframework==0.3.0
python-dotenv==0.19.0
dj-database-url==0.5.0
psycopg2-binary==2.9.1
django-cors-headers==3.7.0
uvicorn==0.15.0
websockets==10.0